Add HostProperties class to hold code dealing with host
